add searchNotes function
