+++ app/(dashboard)/notes/page.tsx
@@ -1,5 +1,6 @@
 import NotesTable, { NoteTableView } from '@/components/NotesTable';
 import { createClient } from '@/utils/supabase/server';
+import { searchNotes } from '@/utils/notes/search-notes';
 import Link from 'next/link';
 import { redirect } from 'next/navigation';
 import { PiCaretLeft } from 'react-icons/pi';
@@ -16,48 +17,7 @@ export default async function Notes({
     return redirect('/');
   }
 
-  let notes: NoteTableView[] | null = [];
-
-  if (tag) {
-    const { data: notesByTags } = await supabase
-      .from('tags')
-      .select(`note_tags ( notes ( id, title, note_tags ( tags ( name ) ) ) )`)
-      .eq('name', tag);
-
-    if (notesByTags?.length) {
-      const notesByTagsMapped: NoteTableView[] = [];
-      notesByTags[0].note_tags.map((note_tag) => {
-        if (note_tag.notes) notesByTagsMapped.push(note_tag.notes);
-      });
-      notes = notesByTagsMapped;
-    }
-  }
-
-  if (search) {
-    const searchQuery = supabase
-      .from('notes')
-      .select(
-        `
-      id,
-      title,
-      note_tags!inner(tags!inner(name))
-      `
-      )
-      .textSearch(`fts_query`, `${search}`, {
-        type: 'websearch',
-        config: 'english',
-      });
-
-    if (tag) {
-      searchQuery.in(
-        'id',
-        notes.map((note) => note.id)
-      );
-    }
-
-    const { data: notesBySearch, error } = await searchQuery;
-    notes = notesBySearch;
-  }
+  const notes = await searchNotes({ supabase, search, tag });
 
   return (
     <div>
